Aci 65:Publishing the resource Added event after  the initial plugin status request  (#69)
* odim 64:Removing association of Endpoints from zoneOfEndpoints during deletion of Endpoints

* aci 65:Publishing resource added the event of fabric during 1st status time

* aci 65:Renaming the function name
